Add support for CommonJS environments
Commonly, Browserify. Can't use AngularFire alongside Browserify at the
moment due to it expecting `window.angular` to exist. A quick check to
see if that's the case means behaviour for the browser stays the same,
else seeing if we can expose the function that binds AngularFire to the
passed in Angular instance.

Used in the browser just as before, but also through Browserify as:

```
var angular = require('angular');
require('angular-fire')(angular);
// this now works fine
var myApp = angular.module("MyApp", ["firebase"]);
```

This isn't possible until angular-fire is published onto npm, but
hopefully that is easily arranged :)